+++ src/utils/physfsrwops.cpp
@@ -121,7 +121,9 @@ static PHYSFSSIZE physfsrwops_read(SDL_RWops *const rw, void *ptr,
 {
     PHYSFS_file *const handle = static_cast<PHYSFS_file *const>(
         rw->hidden.unknown.data1);
-    const PHYSFS_sint64 rc = PHYSFS_read(handle, ptr, size, maxnum);
+    const PHYSFS_sint64 rc = PHYSFS_read(handle, ptr,
+        static_cast<unsigned int>(size),
+        static_cast<unsigned int>(maxnum));
     if (rc != static_cast<PHYSFS_sint64>(maxnum))
     {
         if (!PHYSFS_eof(handle)) /* not EOF? Must be an error. */
@@ -137,7 +139,9 @@ static PHYSFSSIZE physfsrwops_write(SDL_RWops *const rw, const void *ptr,
 {
     PHYSFS_file *const handle = static_cast<PHYSFS_file *const>(
         rw->hidden.unknown.data1);
-    const PHYSFS_sint64 rc = PHYSFS_write(handle, ptr, size, num);
+    const PHYSFS_sint64 rc = PHYSFS_write(handle, ptr,
+        static_cast<unsigned int>(size),
+        static_cast<unsigned int>(num));
     if (rc != static_cast<PHYSFS_sint64>(num))
         SDL_SetError("PhysicsFS error: %s", PHYSFS_getLastError());
